feat: navigation
Created separate components for FooterMenu and MainMenu.
Removed Menu since it's not used anymore.
Simplified homepage sections creation - removed .map
