+++ src/components/FooterMenu/index.ts
@@ -1,0 +1,3 @@
+import FooterMenu from './FooterMenu'
+
+export default FooterMenu
