+++ src/components/MainMenu/index.ts
@@ -1,0 +1,3 @@
+import MainMenu from './MainMenu'
+
+export default MainMenu
