--- src/components/Menu/MenuItemInterface.ts
@@ -1,4 +1,0 @@
-export default interface MenuItemInterface {
-  name: string
-  sectionId: string
-}
