+++ src/components/MenuItem/index.ts
@@ -1,0 +1,3 @@
+import MenuItem from './MenuItem'
+
+export default MenuItem
